requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.5.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2
python-slugify==8.0.1
python-magic==0.4.27
azure-storage-blob==12.17.0
python-magic-bin==0.4.14; sys_platform == 'win32'
pydantic[email]
